move and refactor code
